Add tabs for components nav
